requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
mysqlclient==2.2.4
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
PyJWT==2.9.0
pypng==0.20220715.0
python-decouple==3.8
qrcode==7.4.2
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
